pyproject.toml
[tool.poetry]
name = "trial1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.11, <3.12"
python-dotenv = "^1.0.1"
llama-index = "^0.10.59"
llama-index-embeddings-huggingface = "^0.2.2"
torch = "2.2.2"
chromadb = "^0.5.5"
llama-index-vector-stores-chroma = "^0.1.10"
google-generativeai = ">=0.5.2,<0.6.0"
llama-index-llms-gemini = "^0.1.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
